v2 decision UI backbone
