+++ alcs-frontend/src/app/features/application/application.module.ts
@@ -15,7 +15,8 @@ import { ApplicationMeetingDialogComponent } from './application-meeting/applica
 import { ApplicationMeetingComponent } from './application-meeting/application-meeting.component';
 import { CreateApplicationMeetingDialogComponent } from './application-meeting/create-application-meeting-dialog/create-application-meeting-dialog.component';
 import { ApplicationComponent, childRoutes } from './application.component';
-import { DecisionDialogComponent } from './decision/decision-dialog/decision-dialog.component';
+import { DecisionDialogComponent } from './decision/decision-v1/decision-dialog/decision-dialog.component';
+import { DecisionV1Component } from './decision/decision-v1/decision.component';
 import { DecisionComponent } from './decision/decision.component';
 import { DocumentUploadDialogComponent } from './documents/document-upload-dialog/document-upload-dialog.component';
 import { DocumentsComponent } from './documents/documents.component';
@@ -34,9 +35,10 @@ import { SubdProposalComponent } from './proposal/subd/subd.component';
 import { DecisionMeetingDialogComponent } from './review/decision-meeting-dialog/decision-meeting-dialog.component';
 import { DecisionMeetingComponent } from './review/decision-meeting/decision-meeting.component';
 import { ReviewComponent } from './review/review.component';
-import { StaffJournalComponent } from './staff-journal/staff-journal.component';
-import { StaffJournalNoteComponent } from './staff-journal/staff-journal-note/staff-journal-note.component';
 import { StaffJournalNoteInputComponent } from './staff-journal/staff-journal-note-input/staff-journal-note-input.component';
+import { StaffJournalNoteComponent } from './staff-journal/staff-journal-note/staff-journal-note.component';
+import { StaffJournalComponent } from './staff-journal/staff-journal.component';
+import { DecisionV2Component } from './decision/decision-v2/decision-v2.component';
 
 const routes: Routes = [
   {
@@ -65,7 +67,7 @@ const routes: Routes = [
     ApplicationMeetingComponent,
     ApplicationMeetingDialogComponent,
     CreateApplicationMeetingDialogComponent,
-    DecisionComponent,
+    DecisionV1Component,
     InfoRequestsComponent,
     InfoRequestDialogComponent,
     TimelineComponent,
@@ -84,6 +86,8 @@ const routes: Routes = [
     ProposalComponent,
     NfuProposalComponent,
     SubdProposalComponent,
+    DecisionComponent,
+    DecisionV2Component,
   ],
   imports: [
     SharedModule.forRoot(),
