+++ alcs-frontend/src/app/features/application/decision/decision-v1/decision-dialog/decision-dialog.component.scss
