+++ alcs-frontend/src/app/features/application/decision/decision-v1/decision-dialog/decision-dialog.component.spec.ts
@@ -2,18 +2,18 @@ import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatButtonToggleModule } from '@angular/material/button-toggle';
-import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
+import { MatDialogModule, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { NgSelectModule } from '@ng-select/ng-select';
 import { BehaviorSubject } from 'rxjs';
-import { ApplicationModificationDto } from '../../../../services/application/application-modification/application-modification.dto';
-import { ApplicationModificationService } from '../../../../services/application/application-modification/application-modification.service';
-import { ApplicationDecisionService } from '../../../../services/application/application-decision/application-decision.service';
-import { ApplicationReconsiderationDto } from '../../../../services/application/application-reconsideration/application-reconsideration.dto';
-import { ApplicationReconsiderationService } from '../../../../services/application/application-reconsideration/application-reconsideration.service';
-import { MomentPipe } from '../../../../shared/pipes/moment.pipe';
-import { StartOfDayPipe } from '../../../../shared/pipes/startOfDay.pipe';
+import { ApplicationDecisionService } from '../../../../../services/application/application-decision/application-decision.service';
+import { ApplicationModificationDto } from '../../../../../services/application/application-modification/application-modification.dto';
+import { ApplicationModificationService } from '../../../../../services/application/application-modification/application-modification.service';
+import { ApplicationReconsiderationDto } from '../../../../../services/application/application-reconsideration/application-reconsideration.dto';
+import { ApplicationReconsiderationService } from '../../../../../services/application/application-reconsideration/application-reconsideration.service';
+import { MomentPipe } from '../../../../../shared/pipes/moment.pipe';
+import { StartOfDayPipe } from '../../../../../shared/pipes/startOfDay.pipe';
 import { DecisionDialogComponent } from './decision-dialog.component';
 
 describe('DecisionDialogComponent', () => {
