+++ alcs-frontend/src/app/features/application/decision/decision-v1/decision.component.scss
@@ -1,0 +1,140 @@
+@use '../../../../../styles/colors';
+
+h3 {
+  margin-bottom: 24px !important;
+}
+
+section {
+  margin-bottom: 64px;
+}
+
+.decision-container {
+  position: relative;
+}
+
+.decision {
+  margin: 24px 0;
+  box-shadow: 0 2px 8px 1px rgba(0, 0, 0, 0.25);
+}
+
+.loading-overlay {
+  position: absolute;
+  z-index: 2;
+  background-color: colors.$grey;
+  opacity: 0.4;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.days {
+  display: inline-block;
+  margin-right: 6px;
+
+  .mat-icon {
+    font-size: 19px !important;
+    line-height: 21px !important;
+    width: 19px;
+    vertical-align: middle;
+  }
+}
+
+.post-decisions {
+  padding: 6px 32px;
+  background-color: colors.$accent-color-light;
+  grid-template-columns: 1fr 1fr 100px;
+  display: grid;
+  min-height: 36px;
+  text-transform: uppercase;
+  border-radius: 4px 4px 0 0;
+  box-shadow: -1px 1px 4px rgba(0, 0, 0, 0.25);
+}
+
+.decision-menu {
+  position: absolute;
+  top: 0;
+  right: 0;
+  height: 36px;
+  background: colors.$accent-color;
+  box-shadow: -1px 1px 4px rgba(0, 0, 0, 0.25);
+  border-radius: 0 4px 0 10px;
+
+  button {
+    color: colors.$white;
+    width: 36px;
+    height: 36px;
+    line-height: 36px;
+  }
+
+  mat-icon {
+    position: absolute;
+    top: 8px;
+    left: 8px;
+    font-size: 21px;
+    width: 20px;
+    height: 36px;
+  }
+}
+
+.decision-padding {
+  padding: 18px 32px 32px 32px;
+}
+
+.decision-content {
+  margin-top: 16px;
+  margin-bottom: 32px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 100px;
+  grid-row-gap: 24px;
+
+  .subheading2 {
+    margin-bottom: 6px !important;
+  }
+
+  & > div {
+    font-size: 16px;
+  }
+}
+
+.decision-documents {
+  margin-top: 4px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 100px;
+  grid-row-gap: 4px;
+
+  div {
+    display: flex;
+    align-items: center;
+    font-size: 16px !important;
+  }
+}
+
+.file-actions {
+  margin-left: -12px;
+}
+
+.delete-file-icon {
+  color: colors.$error-color;
+}
+
+.no-decisions {
+  margin-top: 16px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: colors.$grey-light;
+  height: 72px;
+}
+
+.no-files {
+  grid-column: 1/4;
+  margin-top: 16px;
+  padding: 16px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  background-color: colors.$grey-light;
+}
