+++ alcs-frontend/src/app/features/application/decision/decision-v1/decision.component.spec.ts
@@ -1,0 +1,64 @@
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { createMock, DeepMocked } from '@golevelup/ts-jest';
+import { BehaviorSubject } from 'rxjs';
+import { ApplicationDecisionService } from '../../../../services/application/application-decision/application-decision.service';
+import { ApplicationDetailService } from '../../../../services/application/application-detail.service';
+import { ApplicationDto } from '../../../../services/application/application.dto';
+import { ToastService } from '../../../../services/toast/toast.service';
+import { ConfirmationDialogService } from '../../../../shared/confirmation-dialog/confirmation-dialog.service';
+
+import { DecisionV1Component } from './decision.component';
+
+describe('DecisionComponent', () => {
+  let component: DecisionV1Component;
+  let fixture: ComponentFixture<DecisionV1Component>;
+  let mockAppDetailService: DeepMocked<ApplicationDetailService>;
+
+  beforeEach(async () => {
+    mockAppDetailService = createMock();
+    mockAppDetailService.$application = new BehaviorSubject<ApplicationDto | undefined>(undefined);
+
+    await TestBed.configureTestingModule({
+      imports: [MatSnackBarModule],
+      declarations: [DecisionV1Component],
+      providers: [
+        {
+          provide: ApplicationDetailService,
+          useValue: mockAppDetailService,
+        },
+        {
+          provide: ApplicationDecisionService,
+          useValue: {},
+        },
+        {
+          provide: MatDialogRef,
+          useValue: {},
+        },
+        {
+          provide: ConfirmationDialogService,
+          useValue: {},
+        },
+        {
+          provide: ToastService,
+          useValue: {},
+        },
+        {
+          provide: MatDialog,
+          useValue: {},
+        },
+      ],
+      schemas: [NO_ERRORS_SCHEMA],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(DecisionV1Component);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
