+++ alcs-frontend/src/app/features/application/decision/decision-v2/decision-v2.component.html
@@ -1,0 +1,1 @@
+<p>decision-v2 works!</p>
