+++ alcs-frontend/src/app/features/application/decision/decision-v2/decision-v2.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-decision-v2',
+  templateUrl: './decision-v2.component.html',
+  styleUrls: ['./decision-v2.component.scss'],
+})
+export class DecisionV2Component {}
