+++ alcs-frontend/src/app/features/application/decision/decision.component.spec.ts
@@ -1,57 +1,30 @@
-import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { MatDialog, MatDialogRef } from '@angular/material/dialog';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { BehaviorSubject } from 'rxjs';
-import { ApplicationDecisionService } from '../../../services/application/application-decision/application-decision.service';
 import { ApplicationDetailService } from '../../../services/application/application-detail.service';
 import { ApplicationDto } from '../../../services/application/application.dto';
-import { ToastService } from '../../../services/toast/toast.service';
-import { ConfirmationDialogService } from '../../../shared/confirmation-dialog/confirmation-dialog.service';
 
 import { DecisionComponent } from './decision.component';
 
 describe('DecisionComponent', () => {
   let component: DecisionComponent;
   let fixture: ComponentFixture<DecisionComponent>;
-  let mockAppDetailService: DeepMocked<ApplicationDetailService>;
+  let mockApplicationDetailService: DeepMocked<ApplicationDetailService>;
 
   beforeEach(async () => {
-    mockAppDetailService = createMock();
-    mockAppDetailService.$application = new BehaviorSubject<ApplicationDto | undefined>(undefined);
+    mockApplicationDetailService = createMock();
 
     await TestBed.configureTestingModule({
-      imports: [MatSnackBarModule],
-      declarations: [DecisionComponent],
       providers: [
         {
           provide: ApplicationDetailService,
-          useValue: mockAppDetailService,
-        },
-        {
-          provide: ApplicationDecisionService,
-          useValue: {},
-        },
-        {
-          provide: MatDialogRef,
-          useValue: {},
-        },
-        {
-          provide: ConfirmationDialogService,
-          useValue: {},
-        },
-        {
-          provide: ToastService,
-          useValue: {},
-        },
-        {
-          provide: MatDialog,
-          useValue: {},
+          useValue: mockApplicationDetailService,
         },
       ],
-      schemas: [NO_ERRORS_SCHEMA],
+      declarations: [DecisionComponent],
     }).compileComponents();
+
+    mockApplicationDetailService.$application = new BehaviorSubject<ApplicationDto | undefined>(undefined);
 
     fixture = TestBed.createComponent(DecisionComponent);
     component = fixture.componentInstance;
